requirements.txt
boto3
kinesis-python
